# FEAT-#2491: optimized groupby dictionary aggregation