--- a/modin/backends/pandas/query_compiler.py
+++ b/modin/backends/pandas/query_compiler.py
@@ -23,7 +23,7 @@ from pandas.core.dtypes.common import (
 )
 from pandas.core.base import DataError
 from typing import Type, Callable
-from collections.abc import Iterable
+from collections.abc import Iterable, Container
 import warnings
 
 
@@ -2485,7 +2485,6 @@ class PandasQueryCompiler(BaseQueryCompiler):
                 map_fns.append((future_col_name, groupby_reduce_functions[func][0]))
                 reduce_dict[(col, future_col_name)] = groupby_reduce_functions[func][1]
             map_dict[col] = map_fns
-
         return GroupbyReduceFunction.register(map_dict, reduce_dict)(
             query_compiler=self,
             by=by,
@@ -2508,12 +2507,17 @@ class PandasQueryCompiler(BaseQueryCompiler):
         groupby_kwargs,
         drop=False,
     ):
-        def is_reduce_fn(o):
-            if callable(o):
-                return False
-            if isinstance(o, str):
-                o = [o]
-            return all(x in groupby_reduce_functions for x in o)
+        def is_reduce_fn(o, deep_level=0):
+            if not isinstance(o, str) and isinstance(o, Container):
+                assert deep_level == 0 or (
+                    deep_level > 0 and len(o) == 2
+                ), f"Got the renamer with incorrect length, expected 2 got {len(o)}."
+                return (
+                    all(is_reduce_fn(v, deep_level + 1) for v in o)
+                    if deep_level == 0
+                    else is_reduce_fn(o[1], deep_level + 1)
+                )
+            return isinstance(o, str) and o in groupby_reduce_functions
 
         if isinstance(agg_func, dict) and all(
             is_reduce_fn(x) for x in agg_func.values()