--- a/modin/data_management/functions/groupby_function.py
+++ b/modin/data_management/functions/groupby_function.py
@@ -45,7 +45,6 @@ class GroupbyReduceFunction(MapReduceFunction):
         # Set `as_index` to True to track the metadata of the grouping object
         # It is used to make sure that between phases we are constructing the
         # right index and placing columns in the correct order.
-        as_index = groupby_args.get("as_index", True)
         groupby_args["as_index"] = True
         groupby_args["observed"] = True
         if other is not None: