<!--
Thank you for your contribution!
Please review the contributing docs: https://modin.readthedocs.io/en/latest/CONTRIBUTING.html
if you have questions about contributing.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

- [x] commit message follows format outlined [here](https://modin.readthedocs.io/en/latest/CONTRIBUTING.html)
- [x] passes `flake8 modin`
- [x] passes `black --check modin`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #2491 #2512<!-- issue must be created for each patch -->
- [x] tests added and passing

## What was done

- Added [`groupby_reduce_functions` dictionary](https://github.com/dchigarev/modin/blob/021d45e900455eddbbaf1bffa295d50fdfa838ba/modin/data_management/functions/groupby_function.py#L229) of all supported reduction operation in pandas backend, it contains map and reduce functions for every method. That was done to easily translate function name in aggregation dictionary to its proper map and reduce functions.
- Now, when `PandasQueryCompiler.groupby_agg` obtains dictionary as a function, it checks that all function exists in the `groupby_reduce_functions` dictionary, if it is - going to new added [`PandasQueryCompiler._groupby_dict_reduce`](https://github.com/dchigarev/modin/blob/021d45e900455eddbbaf1bffa295d50fdfa838ba/modin/backends/pandas/query_compiler.py#L2459) method, otherwise continues in `groupby_agg`.
- Added [`PandasQueryCompiler._groupby_dict_reduce`](https://github.com/dchigarev/modin/blob/021d45e900455eddbbaf1bffa295d50fdfa838ba/modin/backends/pandas/query_compiler.py#L2459) method to properly dispatch function names to its map+reduce implementations. 
- Refactored [`data_management/functions/groupby_function.py`](https://github.com/dchigarev/modin/blob/dict_agg_op/modin/data_management/functions/groupby_function.py) to reduce the amount of unused code and split map and reduce phase in different methods. Overall logic of `GroupbyReduceFunction` doesn't changed much
- Added partitions enumeration at reduce phase to fix #2512, now we're inserting 'by' columns in case of `as_index=False` only if `partition_idx == 0`
- Added `apply_indices` parameter to [`BaseModinFrame.groupby_reduce`](https://github.com/dchigarev/modin/blob/021d45e900455eddbbaf1bffa295d50fdfa838ba/modin/engines/base/frame/data.py#L1998) method, it allows us to do not touch partitions that do not contains columns to aggregate

## Performance
Number of groups: 100
Times in tables: median of 10 runs
Function: 3 reduction function applied to 3 different columns (`sum, prod, count`)

Since now we're applying 2 functions to partitions instead of 1 (map + reduce), we can see a little regression in dictionary aggregation for small datasets (~200k rows) because we're getting more overhead of partitioning than profit. But when the number of rows becomes bigger (1kk+) we see a significant increase in performance.

<details><summary> 16 cores </summary>


|(200.000, 256)|len(by) == 1|len(by) == 2|len(by) == 6|
|-|-|-|-|
|pandas|0.008s|0.012s|0.023s|
|modin #2534|0.14s|0.17s|0.21s|
|modin master|0.14s|0.15s|0.18s|

|(1.000.000, 256)|len(by) == 1|len(by) == 2|len(by) == 6|
|-|-|-|-|
|pandas|0.029s|0.057s|0.11s|
|modin #2534|0.13s|0.17s|0.24s|
|modin master|0.26s|0.31s|0.41s|


|(10.000.000, 20)|len(by) == 1|len(by) == 2|len(by) == 6|
|-|-|-|-|
|pandas|0.31s|0.6s|1.18s|
|modin #2534|0.37s|0.5s|0.62s|
|modin master|1.29s|1.77s|2.86s|

</details>

<details><summary> 112 cores </summary>


|(200.000, 256)|len(by) == 1|len(by) == 2|len(by) == 6|
|-|-|-|-|
|pandas|0.008s|0.012s|0.023s|
|modin #2534|0.36s|0.5s|0.7s|
|modin master|0.4s|0.47s|0.63s|

|(1.000.000, 256)|len(by) == 1|len(by) == 2|len(by) == 6|
|-|-|-|-|
|pandas|0.029s|0.057s|0.11s|
|modin #2534|0.37s|0.49s|0.69s|
|modin master|0.57s|0.64s|0.81s|


|(10.000.000, 20)|len(by) == 1|len(by) == 2|len(by) == 6|
|-|-|-|-|
|pandas|0.31s|0.6s|1.18s|
|modin #2534|0.41s|0.59s|0.78s|
|modin master|1.73s|2.06s|3.16s|

</details>

<details><summary> Script to measure </summary>


```python
import numpy as np
import os
import pandas
from modin.utils import try_cast_to_pandas
from timeit import default_timer as timer


def generate_data_file(
    nrows=200000,
    ncols=256,
    ngroups=100,
    filename=None,
    force=False,
    rand_low=-100,
    rang_high=100,
):
    if filename is None:
        filename = os.path.abspath(
            f"int_dataset-{nrows},{ncols},{rand_low},{rang_high},{ngroups}.csv"
        )

    print("generating", filename)

    if os.path.exists(filename) and not force:
        return filename

    data = {
        f"col{i}":
        # Generates ngroups for groupby in each column
        list(np.random.randint(rand_low, rang_high, ngroups)) * (nrows // ngroups)
        for i in np.arange(ncols)
    }
    print("dict generated!")
    df = pandas.DataFrame(data)
    print("dataframe created!")
    df.to_csv(filename)
    print("csv ready!")
    return filename


def measure_fn(fn, n):
    times = []
    for _ in range(n):
        t1 = timer()
        fn()
        times.append(timer() - t1)
    return np.min(times), np.median(times), times


def measure_time(df, by, n, **kwargs):
    print(f"\n======= Testing {kwargs['operation'][0]} =======")
    # Printing parameters:
    kwargs.update(
        {"Frame shape": df.shape, "len(by)": len(by), "ncores": pd.DEFAULT_NPARTITIONS}
    )
    op = kwargs.pop("operation")[1]

    print("Parameters:")
    for k, v in kwargs.items():
        print(f"\t{k}: {v}")

    as_index = kwargs["as_index"]

    def measure():
        md_res = op(df.groupby(by, as_index=as_index))
        # shape to trigger materialization
        md_res.shape

    min_t, median_t, times = measure_fn(measure, n)

    print(f"Min time: {min_t}s | Median time: {median_t}s")


if __name__ == "__main__":
    import modin.pandas as pd
    from modin.pandas.test.utils import df_equals

    pd.DEFAULT_NPARTITIONS = 16

    nrows = [200_000, 1_000_000]

    by_cols = [
        ["col0"],
        ["col0", "col1"],
        ["col0", "col1", "col2", "col3", "col4", "col5"],
    ]
    operation = [
        ("agg", lambda grp: grp.agg({"col6": "sum", "col7": "prod", "col8": "count"})),
    ]
    as_index = [True]
    ngroups = 100

    for nrows_ in nrows:
        fname = generate_data_file(nrows_, ngroups=ngroups)
        md_df = pd.read_csv(fname)
        for op in operation:
            for by in by_cols:
                for as_ind in as_index:
                    measure_time(
                        md_df,
                        by,
                        operation=op,
                        ngroups=ngroups,
                        as_index=as_ind,
                        n=10,
                    )
```


</details>

## ASV runs:
As described above we can see some degradation for small frames, but when the frame becomes bigger (1kk+ rows) performance increased:
<details><summary> 16 cores </summary>


```
MODIN_CPUS=16 asv continuous src/master HEAD -b TimeGroupBy --launch-method=spawn -a number=5
       before           after         ratio
     [f98a7b3a]       [830d2ea3]
     <master>        <dict_agg_op>
+      56.1±0.4ms       80.2±0.7ms     1.43  benchmarks.TimeGroupBy.time_groupby_dictionary_reduction('modin', 'int', (2000, 100))

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.


MODIN_CPUS=16 MODIN_TEST_DATASET_SIZE=Big asv continuous src/master HEAD -b TimeGroupBy --launch-method=spawn -a number=5
       before           after         ratio
     [f98a7b3a]       [830d2ea3]
     <master>        <dict_agg_op>
+        99.9±2ms          127±2ms     1.27  benchmarks.TimeGroupBy.time_groupby_dictionary_reduction('modin', 'int', (5000, 5000))
-         167±4ms          127±3ms     0.76  benchmarks.TimeGroupBy.time_groupby_dictionary_reduction('modin', 'int', (10, 1000000))
SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```


</details>


<details><summary> 112 cores </summary>



```
MODIN_CPUS=112 asv continuous src/master HEAD -b TimeGroupBy --launch-method=spawn -a number=5
       before           after         ratio
     [f98a7b3a]       [830d2ea3]
     <master>        <dict_agg_op>
+      56.9±0.4ms          150±2ms     2.64  benchmarks.TimeGroupBy.time_groupby_dictionary_reduction('modin', 'int', (2000, 100))

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.


MODIN_CPUS=112 MODIN_TEST_DATASET_SIZE=Big asv continuous src/master HEAD -b TimeGroupBy --launch-method=spawn -a number=5
       before           after         ratio
     [f98a7b3a]       [830d2ea3]
     <master>        <dict_agg_op>
-         482±8ms          368±8ms     0.76  benchmarks.TimeGroupBy.time_groupby_dictionary_reduction('modin', 'int', (10, 1000000))

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```


</details>