@anmyachev 2020-12-16
@dchigarev TimeGroupBy suite does not contain perf tests with dictionary aggregations, can you add the ones?