@dchigarev 2020-12-16
@anmyachev added new benchmark suits, also updated [PR comment](https://github.com/modin-project/modin/pull/2534#issue-539468820) with new ASV results